Moved AuthorizingPaymentURLVerifier to root package.
